requirements.txt
numpy==1.26.4
mediapipe==0.10.9
opencv-python==4.9.0.80
ffmpeg-python==0.2.0
tqdm==4.66.2
Flask==3.0.2
Flask-RESTful==0.3.10
Pillow==9.3.0